requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
numpy
scikit-learn
